% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asME.R
\name{asME}
\alias{asME}
\title{Make MoleculeExperiment with transcript data already read into R.}
\usage{
asME(transcripts_df, n_samples = 1, technology = NULL)
}
\arguments{
\item{transcripts_df}{A data.frame containing the transcripts information
for one or more samples.}

\item{n_samples}{Integer specifying the number of samples in the
transcripts_df}

\item{technology}{Character string specifying the technology used for the
data. Can be "xenium", "cosmx" or "merscope".}
}
\description{
This function is an alternative to readXenium, for when transcript file/s
of interest that are already read in R need to be converted to an ME obj.
}
