% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeToMEList.R
\name{dataframeToMEList}
\alias{dataframeToMEList}
\title{Convert a transcripts or boundaries file to the ME list format}
\usage{
dataframeToMEList(df, list_id)
}
\arguments{
\item{df}{A data.frame containing the transcripts information or the
boundaries information. NOTE: this data.frame should, at a minimum, have the
following 4 columns: sample_id, feature_name, x_location and y_location.}

\item{list_id}{Character string specifying the header name of the list. This
is important to store different versions of the transcripts information in a
MoleculeExperiment object downstream.}
}
\description{
The goal of this function is to prepare transcripts and boundaries files for
input to a MoleculeExperiment object.
}
