% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.generateBPParam}
\alias{.generateBPParam}
\title{Utility function to generate BPPARAM object.}
\usage{
.generateBPParam(cores = 1)
}
\arguments{
\item{cores}{Desired number of cores for BPPARAM object.}
}
\value{
A BPPPARAM object.
}
\description{
Utility function to generate BPPARAM object.
}
