% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/MoleculeExperiment.R
\docType{class}
\name{MoleculeExperiment-class}
\alias{MoleculeExperiment-class}
\alias{MoleculeExperiment}
\title{MoleculeExperiment class: An S4 class container to store imaging-based
spatial transcriptomics data.}
\usage{
MoleculeExperiment(molecules)
}
\arguments{
\item{molecules}{Detected transcripts information in a list format, as is
generated with the readMolecules() function.}
}
\description{
This class enables the analysis of imaging-based ST data at the molecule
level, and standardises data across vendors, which hopefully facilitates
ST data integration and comparison.
}
\section{Slots}{

\describe{
\item{\code{molecules}}{}
}}

\section{Creating an ME object}{
NA
}

\section{Methods}{
NA
}

