% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_readBoundaries.R
\name{old_readBoundaries}
\alias{old_readBoundaries}
\title{Read segmentation boundary files into a MoleculeExperiment object.}
\usage{
old_readBoundaries(
  me,
  boundaries_mode = NULL,
  data_dir,
  pattern = NULL,
  n_samples = NULL
)
}
\arguments{
\item{me}{A MoleculeExperiment object. Note: a MoleculeExperiment object can
only be created with a filled molecules slot.}

\item{boundaries_mode}{Character string naming the list of the boundaries
slot that will contain the imported boundary information. This is important
to recognise the imported boundaries later on. For example, one might want
to store cell boundaries in a list entry named "cells", and the nuclei
boundaries in a list entry named "nuclei". These different entries can be
accessed with the boundaries() accessor (see boundaries() for more
information).}

\item{data_dir}{Character string indicating path to the parent directory
containing the boundary files for the sample/s.}

\item{pattern}{Character string specifying a pattern unique to the name of
the csv boundary file/s of interest. E.g., "cell_boundaries.csv"}

\item{n_samples}{Integer indicating the number of samples from which to read
boundary files.}
}
\description{
This function reads in boundary files of interest, standardises the data
format, and adds the data to the "boundaries" slot of a specified
MoleculeExperiment object. NOTE: currently, this function works with csv
segmentation files with the format given by the segmentation method of
10X genomics. I.e., a csv file with the following columns and column order:
"cell_id", "vertex_x" and "vertex_y".
}
