% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBoudaries}
\alias{plotBoudaries}
\title{plotBoudaries}
\usage{
plotBoudaries(me)
}
\arguments{
\item{me}{A molecule experiment with boundaries.}
}
\value{
A boundary plot.
}
\description{
Plots the molucules and boundaries of a ME object.
}
\examples{
repo_dir <- system.file("extdata", package = "MoleculeExperiment")

me <- readXenium(
    repo_dir,
    keepCols = "essential"
)
me

nuclei_ls <- readBoundaries(
    repo_dir,
    pattern = "nucleus_boundaries.csv",
    segmentIDCol = "cell_id",
    xCol = "vertex_x",
    yCol = "vertex_y",
    keepCols = "essential",
    boundariesAssay = "nucleus",
    scaleFactorVector = 1
)

boundaries(me, "nucleus") <- nuclei_ls

plotBoudaries(me)

}
