% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBoundaries.R
\name{readBoundaries}
\alias{readBoundaries}
\title{#TODO Description of function.}
\usage{
readBoundaries(
  data_dir,
  pattern = NULL,
  n_samples = NULL,
  segment_id_col = NULL,
  x_col = NULL,
  y_col = NULL,
  keep_cols = "essential",
  boundaries_assay = NULL,
  scale_factor_vector = 1
)
}
\arguments{
\item{data_dir}{Path of the directory containing the data.
#TODO @param pattern DESCRIPTION. Defaults to NULL.
#TODO @param n_samples DESCRIPTION. Defaults to NULL.
#TODO @param segment_id_col DESCRIPTION. Defaults to NULL.
#TODO @param x_col DESCRIPTION. Defaults to NULL.
#TODO @param y_col DESCRIPTION. Defaults to NULL.
#TODO @param keep_cols DESCRIPTION. Defaults to "essential".
#TODO @param boundaries_assay DESCRIPTION. Defaults to NULL.}

\item{scale_factor_vector}{Vector containing the scale factor/s with which to
change the coordinate data from pixel to micron. It can be either a single
integer, or multiple scale factors for the different samples. The default
value is 1.}
}
\description{
#TODO Description of function.
}
