% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCosmx.R
\name{readCosmx}
\alias{readCosmx}
\title{Read in Cosmx data (Nanostring) as an ME object.}
\usage{
readCosmx(data_dir, n_samples = NULL, keep_cols = "essential")
}
\description{
This function is a wrapper around the readMolecules function. Note that it
can currently only create a simple ME object with the molecules slot filled.
Boundary information is not handled yet.
}
\details{
#TODO @param data_dir DESCRIPTION.
#TODO @param n_samples DESCRIPTION. Defaults to NULL.
#TODO @param keep_cols DESCRIPTION. Defaults to "essential".
}
