% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMerscope.R
\name{readMerscope}
\alias{readMerscope}
\title{Read in Merscope data to an ME object}
\usage{
readMerscope(data_dir, n_samples = 1, keep_cols = "essential")
}
\description{
Reads in Merscope data (Vizgen) from a directory, and standardises it into
a MoleculeExperiment object. It is essentially a wrapper around the function
readMolecules(). Note that this function can currently only create a simple
ME object with the molecules slot filled. Boundary information cannot be
handled yet.
}
