% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMolecules.R
\name{readMolecules}
\alias{readMolecules}
\title{Read and standardise the detected transcripts file/s into a
MoleculeExperiment object}
\usage{
readMolecules(
  data_dir,
  technology = "xenium",
  n_samples = 1,
  essential_cols = c("feature_name", "x_location", "y_location"),
  keep_all_cols = FALSE
)
}
\arguments{
\item{data_dir}{String specifying directory with the file/s with detected
transcripts for different runs/samples.}

\item{technology}{String specifying whether input data was generated with
"xenium" (10X Genomics), "cosmx smi" (Nanostring), or "merscope" (Vizgen).}

\item{n_samples}{Integer specifying number of samples to be read.}

\item{cols}{Vector of characters specifying the names of the columns
containing the gene names, x locations and y locations.}
}
\value{
A standardised detected transcripts file across different
imaging-based spatial transcriptomics technologies. This file can be used
as input for creating a MoleculeExperiment object.
}
\description{
Read and standardise the detected transcripts file/s into a
MoleculeExperiment object
}
