% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSegMask.R
\name{readSegMask}
\alias{readSegMask}
\title{Read a segmentation mask}
\usage{
readSegMask(
  extent,
  path = NULL,
  image = NULL,
  assayName = "cell",
  background_value = NULL,
  sample_id = NULL
)
}
\arguments{
\item{extent}{The extent of the loaded segmentation mask in micrometers.
Used to align the mask with the transcripts. This must be of the form
c(xmin, xmax, ymin, ymax).}

\item{path}{The path of the segmenation mask, Default: NULL}

\item{image}{The loaded image object, Default: NULL}

\item{assayName}{The name of the segmentation (e.g. cell, or nucleus),
Default: 'cell'}

\item{background_value}{The value corrisponding to the backgorund in the
segmentation, Default: NULL}

\item{sample_id}{What the sample should be named, Default: NULL}
}
\value{
A boundaries object.
}
\description{
Reads a segmentation mask TIFF and transforms it into a
ME boundaries object. One must provide either the path or the loaded
image object.
}
\examples{
\dontrun{
if (interactive()) {
    # EXAMPLE1
}
}
}
