% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoleculeExperiment.R
\name{simple_me}
\alias{simple_me}
\title{S4 class for imaging-based ST data summarisation and visualisation}
\usage{
simple_me(molecules)
}
\arguments{
\item{mdf}{molecule data frame from ST experiment. Output from readMolecules().}
}
\description{
The ME obj enables standardised data storage and data querying for
imaging based ST data.
}
