% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_me}
\alias{small_me}
\title{A subsetted Xenium dataset include for demostration purposes.}
\format{
An object of class \code{MoleculeExperiment} of length 1.
}
\usage{
small_me
}
\description{
A subsetted Xenium dataset include for demostration purposes.
}
\keyword{datasets}
