% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\docType{methods}
\name{summarization}
\alias{summarization}
\alias{strMolecules}
\alias{strBoundaries}
\alias{nFeatures}
\alias{nTranscripts}
\alias{strMolecules,MoleculeExperiment-method}
\alias{strBoundaries,MoleculeExperiment-method}
\alias{nFeatures,MoleculeExperiment-method}
\alias{nTranscripts,MoleculeExperiment-method}
\title{Summarization methods to get insights into a MoleculeExperiment object}
\usage{
\S4method{strMolecules}{MoleculeExperiment}(object)

\S4method{strBoundaries}{MoleculeExperiment}(object)

\S4method{nFeatures}{MoleculeExperiment}(object, assay_name = "detected", per_sample = FALSE)

\S4method{nTranscripts}{MoleculeExperiment}(object, assay_name = "detected", per_sample = FALSE)
}
\arguments{
\item{object}{Name of MoleculeExperiment object of interest.}

\item{assay_name}{Character string specifying the name of the assay from
which to view a summary of the contents.}

\item{per_sample}{Logical value specifying whether or not to summarize the
information per sample.}
}
\value{
A MoleculeExperiment object summary.
}
\description{
The following methods are useful to get quick view of the contents in a
MoleculeExperiment object.
For example, strMolecules and strBoundaries summarise the large nested ME
list of lists in the molecules and boundaries slots.
nFeatures and nTranscripts get the numbers of features or transcripts,
respectively. They can do so across all samples, or per sample.
}
\examples{
# get example data
repo_dir <- system.file("extdata", package = "MoleculeExperiment")
me <- readXenium(repo_dir,
                  n_samples = 2,
                  keep_cols = "essential",
                  add_boundaries = "cell")

strMolecules(me)
strBoundaries(me)

nFeatures(me)
nFeatures(me, per_sample = TRUE)

nTranscripts(me)
nTranscripts(me, per_sample = TRUE)
}
