% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeToMEList.R
\name{dataframeToMEList}
\alias{dataframeToMEList}
\title{Convert a transcripts or boundaries file to the ME list format}
\usage{
dataframeToMEList(
  df,
  sample_col,
  factor_col,
  x_coord_col,
  y_coord_col,
  keep_cols = "essential"
)
}
\arguments{
\item{df}{A data.frame containing the transcripts information or the
boundaries information. NOTE: this data.frame should, at a minimum, have the
following 4 columns: sample_id, feature_name, x_location and y_location.}

\item{sample_col}{Character string specifying the name of the column with the
sample id.}

\item{factor_col}{Character string specifying the name of the column with the
factors with which to group the data in the lists. When working with
molecules, this column would be e.g., "feature_name" in xenium. When working
with boundaries, this column would be e.g., "cell_id" in xenium.}

\item{x_coord_col}{Character string specifying the name of the column with
global x coordinates.}

\item{y_coord_col}{Character string specifying the name of the column with
global y coordinates.}

\item{keep_cols}{Character string which can be either "essential" or "all".
If "essential", the function will only work with the x and y location
information.}
}
\description{
The goal of this function is to prepare transcripts and boundaries files for
input to a MoleculeExperiment object.
}
