% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotting functions}
\alias{plotting functions}
\alias{ggplot_me}
\alias{geom_point_me}
\alias{geom_polygon_me}
\title{Plotting functions for SpatialUtils}
\arguments{
\item{me}{MoleculeExperiment object.}

\item{assayName}{Character string specifying name of assay from which to get
data.}

\item{byColour}{Character string specifying the column name to colour by.}

\item{byFill}{Character string specifying the column name to fill by.}
}
\value{
A plot with transcripts and/or segmentation information for imaging
based spatial transcriptomics data.
}
\description{
A set of ggplot functions to build customized plots for imaging based spatial
transcriptomics data.
}
\examples{
repoDir <- system.file("extdata", package = "MoleculeExperiment")
me <- readXenium(repoDir,
                  keepCols = "essential",
                  addBoundaries = c("cell", "nucleus"))

g = ggplot_me() +
        geom_polygon_me(me, byFill = "segment_id", colour = "black") +
        geom_point_me(me, byColour = "feature_name", size = 0.1) +
        geom_polygon_me(me, assay_name = "nucleus", fill = NA, colour = "red")
g
}
