% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCosmx.R
\name{readCosmx}
\alias{readCosmx}
\title{Read in Cosmx data (Nanostring) as an ME object.}
\usage{
readCosmx(data_dir, n_samples = NULL, keep_cols = "essential")
}
\arguments{
\item{data_dir}{Character string specifying the directory with the Cosmx
output files.}

\item{n_samples}{Integer indicating the number of samples. Defaults to NULL.}

\item{keep_cols}{Character string specifying which columns to keep.
Defaults to "essential". The other option is to select "all", or custom
columns by specifying their names in a vector.}
}
\description{
This function is a wrapper around the readMolecules function. Note that it
can currently only create a simple ME object with the molecules slot filled.
Boundary information is not handled yet.
}
