% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCosmx.R
\name{readCosmx}
\alias{readCosmx}
\title{Read in Cosmx data (Nanostring) as an ME object.}
\usage{
readCosmx(dataDir, keepCols = "essential")
}
\arguments{
\item{dataDir}{Character string specifying the directory with the Cosmx
output files.}

\item{keepCols}{Character string specifying which columns to keep.
Defaults to "essential". The other option is to select "all", or custom
columns by specifying their names in a vector.}
}
\value{
A MoleculeExperiment object
}
\description{
This function is a wrapper around the readMolecules function. Note that it
can currently only create a simple ME object with the molecules slot filled.
Boundary information is not handled yet.
}
