% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXenium.R
\name{readXenium}
\alias{readXenium}
\title{Read in Xenium data into a MoleculeExperiment object}
\usage{
readXenium(
  data_dir,
  n_samples = NULL,
  keep_cols = "essential",
  add_boundaries = "cell"
)
}
\arguments{
\item{data_dir}{Inherit docs from readMolecules HERE}

\item{add_boundaries}{Vector with which to specify the names of the boundary
assays to be added to the me object. Can be "cell", "nucleus", both, or NULL.
The latter will lead to the creation of a simple ME object with just the
molecules slot filled.}
}
\description{
Function to read in, and standardise, Xenium output into an ME object.
Detected transcript files are required. Additionally, it is also possible
to read in boundary files ("cell", "nuclei", or both). When boundaries are
read in, the function also reads in the scale factor information from the
experiment.xenium json files for each sample. This function is essentially
a wrapper around readMolecules and readBoundaries functions.
}
