% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readXenium.R
\name{readXenium}
\alias{readXenium}
\title{Read xenium data into a MoleculeExperiment object}
\usage{
readXenium(
  data_dir,
  n_samples = NULL,
  keep_cols = "essential",
  add_boundaries = TRUE,
  boundaries_mode = "cells"
)
}
\description{
Read xenium data into a MoleculeExperiment object
}
